project1_site/requirements.txt
codegen==1.0
Django==2.2.1
djangorestframework==3.9.4
patterns==0.3
pytz==2019.1
sqlparse==0.3.0